+++ NEWS.md
@@ -1,5 +1,9 @@
 # pRoloc 1.9
 
+## CHANGES IN VERSION 1.9.5
+
+- (nothing yet)
+
 ## CHANGES IN VERSION 1.9.4
 
 - fixed remap=FALSE bug in compare app <2017-01-12 Thu>